Add test VF with 4 glyph subset.
Compiling the all whole VF it's not "quick" :) Compiling takes a couple of hours.
That' why I currently compile on a glyphset of only 4 glyphs.

There's a local folder _develop/ that I use to compile subsets for testing, which is in gitignore, not to pollute the repo too much.
I added the latest test VF in fonts/BitcountMono_Double2-TEST-VF.ttf
The script that created it is in develop/COLRv1/test010/build10.py

Type "abce" in FontGoggles. Currently /e contains the canvas that shows the pixel surroundings for both layers.

Two things I need to add here:
• Add the other main 4 Bitcount axes (slnt, wght, OPEN and SHPE). I'll do that in the weekend for this test font.
• I would like to add more shapes to the layer than just concentric gradients. Please some assistants for that code. The current setup of the layers is in that script build10.py