Add non-linear mapping to the OPEN (splitting the pixels in 2 quadrants) axis in the Template.designspace.
Most of the usage is in the small openings, so now there is mapping on the OPEN axis as:
            <map input="500" output="125" />
            <map input="750" output="250" />
1/8 of the 1000 open is done in the first 1/2 of the axis/
Next 1/4 is done in the 1/4 of the axis
Next 1/2 is done in 1/8 of the axis = most of the effect is done here, but not really visible because of the distance.
This gives a semi-curve mapping in only 2 extra points.
This mapping also makes random selections such as Axis-Praxis more likely to select a not-too-open distance.