Add some first JS animations to show the working of the color layers. Run Test-003 on Chrome
(The red dot is there only for testing the sequence now).
Instead of my first "random" floating through the rwo design spaces, this animation is using a table of interesting places, stays there for some time and then animates to the next one.
There will be a longer list of interesting places, as in a catalogue, and their coordinates will be shown.